Add pom file for tachyon.shell and move shell implementation files to shell/ from core/
