+++ shell/src/main/java/tachyon/shell/Utils.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.command;
+package tachyon.shell;
 
 import java.io.IOException;
